Added handlerPath parameter to tray config
Tray can now be told from where to load the features, which is useful in
ARR scenarios (your application is only handling example.com/path*).
Remember to update the path of the FeatureBee handler in the web.config
to path/featurebee.axd + ignoring the route in RouteConfig.
Alternatively you can rewrite requests to path/featurebee.axd to
featurebee.axd in your web.config.